aws-lambda-tesseract-layer-5.3.2 copy/example/serverless/requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
packaging==23.1; python_version >= '3.7'
pillow==10.0.0
pytesseract==0.3.10
requests==2.31.0
urllib3==1.26.7; python_version >= '3.7'
opencv-python-headless
matplotlib
numpy